refactor: move useControllableReducer to its own folder.
