+++ packages/mui-base/src/index.js
@@ -8,6 +8,8 @@ export * from './useFormControlUnstyledContext';
 
 export * from './useInput';
 
+export * from './useControllableReducer';
+
 export { default as BadgeUnstyled } from './BadgeUnstyled';
 export * from './BadgeUnstyled';
