+++ packages/mui-base/src/useControllableReducer/index.ts
@@ -1,0 +1,3 @@
+export { default as useControllableReducer } from './useControllableReducer';
+export { default } from './useControllableReducer';
+export * from './useControllableReducer';
